refactor: extract auth/session boundary into lib/auth-session.js
Move session configuration, passport initialization/strategies, auth helpers,
and login/logout routes from server.js (1825 lines) into a focused module.

This is the first step in splitting server.js by boundary per issue #540.
The auth-session module handles:
  - Session config with Redis/MemoryStore support
  - Passport local + OIDC strategy setup
  - Auth helper functions (getReturnTo, requestWantsJson, etc.)
  - Mobile auth token handoff
  - Login/logout routes (/login, /auth/oidc, /logout, etc.)

server.js reduced from 1825 to ~1360 lines.

Fixes #540
